Update docs, ribbon material translation, and tests for T1 completion
