+++ custom_components/niimbot/translations/en.json
@@ -173,7 +173,7 @@
                 "name": "Ribbon SKU"
             },
             "ribbon_type": {
-                "name": "Ribbon Type"
+                "name": "Ribbon Material"
             },
             "ribbon_tag_uuid": {
                 "name": "Ribbon Tag UUID"
